+++ simplebot_manga/manga_api/__init__.py
@@ -4,6 +4,7 @@ from typing import Dict, List, Optional
 
 from .asurascans import AsuraScans
 from .base import Language, Site
+from .heavenmanga import HeavenManga
 from .kissmanga import KissManga
 from .mangabuddy import MangaBuddy
 from .manganelo import Manganelo
@@ -21,6 +22,7 @@ sites = [
     MangaTigre(),
     ManhuaKO(),
     MangaBuddy(),
+    HeavenManga(),
 ]
 
 lang2sites: Dict[Language, List[Site]] = {}