+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Changed cache configuration to avoid discarding data too soon
 
+### Fixed
+
+- Fixed Spanish "Nine Manga" downloader
+
 ## [v0.2.0]
 
 ### Added